refactor(admin): split ShopsPage god file (1018 lines) into shops/ modules
- Extract usePasswordStrength.ts, useShopMembers.ts, useAddMember.ts, useEditShop.ts hooks
- Extract EditShopModal.tsx, MemberPasswordField.tsx components
- Move main component to shops/ShopsPage.tsx (272 lines, down from 1018)
- Add page-nav handlers to useAuditLogs and useShopMembers to reduce inline callbacks
- Create shops/labels.ts for shopRoleLabel / shopAuditActionLabel
- Create shops/index.tsx re-export
- Replace pages/admin/ShopsPage.tsx with 2-line re-export for router compat
- No behavior changes; TypeScript passes clean
