+++ frontend/src/pages/admin/shops/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './ShopsPage';
